+++ CHANGELOG.md
@@ -1,11 +1,29 @@
 # Change Log
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased][unreleased]
+## [Unreleased Changes][unreleased]
+## Added
+- **MigrateInstance** Artisan command
+- **MigrateInstanceTest** unit test for **MigrateInstance** Artisan command
+- Add package `dreamfactory/dfe-capsule` for instance encapsulation
+
+## Updated
+- Updated `.env-dist` with new default directories for blueprint and migration services
+
+## [1.0.2][v1.0.2]
+### Changed
+- **ImportJob** base class incorrect
+
+## [1.0.1][v1.0.1]
+### Removed
+- Cache files removed from the repository
+
+## [1.0.0][v1.0.0]
 ### Added
-- README file added
-- CHANGELOG file added
+- Initial Release
 - Full suite of command line tests
 
-[unreleased]: https://github.com/dreamfactorysoftware/dfe-console/compare/develop...HEAD
-gith
+[v1.0.2]: https://github.com/dreamfactorysoftware/dfe-console/compare/1.0.2...1.0.1
+[v1.0.1]: https://github.com/dreamfactorysoftware/dfe-console/compare/1.0.1...1.0.0
+[v1.0.0]: https://github.com/dreamfactorysoftware/dfe-console/compare/1.0.0...master
+[unreleased]: https://github.com/dreamfactorysoftware/dfe-console/compare/master...develop